app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Intermediate service for Buisiness process management tool"
authors = ["Nikolay Marusov <n-marusov@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
tenacity = "^6.3.1"
pydantic = "^1.7.3"
gunicorn = "^20.0.4"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
sentry-sdk = "^0.19.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
mypy = "^0.800"
black = "^20.8b1"
autoflake = "^1.4"
pytest-cov = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
